refactor: improve holding logo validation and fallback handling
