+++ src/features/transactionTable/components/columns/columns.tsx
@@ -1,10 +1,11 @@
 import { ColumnDef } from '@tanstack/react-table';
-import { TableColumnHeader } from '@/components/TableColumnHeader';
+import { TableColumnHeader } from '@/components/DataTables/TableColumnHeader';
 import { TransactionTableData } from '@/components/AddTransactionForm/AddTransactionForm';
 import { Button } from '@/components/ui/button';
 import { Edit as EditIcon } from 'lucide-react';
 import { AddTransactionModal } from '@/components/AddTransactionModal';
 import { DeleteButton } from './DeleteButton';
+import { HoldingCell } from '@/components/DataTables/HoldingCell';
 
 export const columns: ColumnDef<TransactionTableData>[] = [
   {
@@ -41,24 +42,7 @@ export const columns: ColumnDef<TransactionTableData>[] = [
     cell: ({ row }) => {
       const holding = row.getValue<TransactionTableData['holding']>('holding');
 
-      return (
-        <div className="flex items-center space-x-4 text-left">
-          <div
-            className="flex items-center justify-center rounded-full bg-primary"
-            style={{ width: '3rem', height: '3rem', flexShrink: 0 }}
-          >
-            <img
-              src={holding.holdingIcon}
-              alt={holding.holdingSymbol}
-              className="h-8 w-8 object-cover"
-            />
-          </div>
-          <div>
-            <div className="font-medium">{holding.holdingSymbol}</div>
-            <div>{holding.holdingName}</div>
-          </div>
-        </div>
-      );
+      return <HoldingCell holding={holding} />;
     },
   },
   {
